Rename unit test files
